Spending credits on likes is completely functional
Liking and superliking jobs or jobseekers decreases credits and performs the necessary actions (like and match) when those fields don't already contain the supplied data. Response contains changes and a Boolean for whether a match was achieved, which will be useful for triggering a new match event on the front end. Skips no longer decrease credits, and neither do non-unique likes/superlikes.