Add picture upload and rendering to jobseeker profile
Seekers can now upload a picture, or replace as needed. Placeholder image is in local filesystem as a fallback for when no uploaded image exists. Image url is stored on the seeker document in the database, delivered through GET to `/profile`, and mapped to props via state.user.profile.imgUrl.