Flow on messaging and browsing optimized
Modals are popping up in response to all errors, as well as a few successful events like matching and successful message sending. Empty job/seeker candidates now renders a composed message with a title, paragraph body, and a button with the appropriate call to action. Posting messages now instantly updates the both the messageHistory with the recipient and the conversations "sidebar" (currently rendering on top of the container).